+++ R/iwa.R
@@ -25,7 +25,7 @@
 #' }
 #'
 #' @examples
-#' \dontrun{
+#' \donttest{
 #' # All four IWA variables for one HUC12, full historical record
 #' df <- nwaa_iwa(
 #'   variable_ids = c("sui", "availab", "strflow", "consum"),